+++ .eleventyignore
@@ -1,0 +1,3 @@
+README.md
+passthrough/**
+node_modules/**
