+++ .github/workflows/static.yml
@@ -24,99 +24,96 @@ concurrency:
 jobs:
   build-dev:
     runs-on: ubuntu-latest
+    concurrency:
+      group: ${{ github.workflow }}-${{ github.ref }}
     steps:
       - name: Checkout
         uses: actions/checkout@v3
-      - name: Make destination directory for compiled CSS
-        run: mkdir -vp ./css
-      - name: Sass Build
-        uses: gha-utilities/sass-build@v0.5.1
+
+      - name: Setup Node
+        uses: actions/setup-node@v3
         with:
-          source: |
-            scss/index.scss
-            scss/menyow.scss
-            scss/styles.scss
-            scss/website.scss
-            scss/terrain.scss
-            scss/hockey.scss
-          destination: |
-            css/index.css
-            css/menyow.css
-            css/styles.css
-            css/website.css
-            css/terrain.css
-            css/hockey.css
-          outputStyle: compressed
-          sourceComments: false
-          sourceMap: false
+          node-version: '18'
+
+      - name: Persist npm cache
+        uses: actions/cache@v3
+        with:
+          path: ~/.npm
+          key: ${{ runner.os }}-node-${{ hashFiles('**/package.json') }}
+
+      - name: Persist Eleventy .cache
+        uses: actions/cache@v3
+        with:
+          path: ./.cache
+          key: ${{ runner.os }}-eleventy-fetch-cache
+
+      - run: npm install
+      - run: npm run build
+      
       - name: Upload artifact
         uses: actions/upload-artifact@v3
         with:
           # Upload entire repository
-          path: '.'
+          path: '_site'
           name: dev
           
-  build-scra:
-    if: "!contains(github.event.commits[0].message, '[skip ci]')"
-    needs: build-dev
-    runs-on: ubuntu-latest
-    environment:
-      name: github-pages
+  # build-scra:
+  #   if: "!contains(github.event.commits[0].message, '[skip ci]')"
+  #   needs: build-dev
+  #   runs-on: ubuntu-latest
+  #   environment:
+  #     name: github-pages
   
-    steps:
-      - name: Checkout SCRA
-        uses: actions/checkout@v3
-        with:
-          repository: ASoftGlow/scra
-          token: ${{ secrets.GH_SCRA_PAT }}
-      - name: Setup Node
-        uses: actions/setup-node@v3
-        with:
-          node-version: "lts/*"
-          cache: npm
-      - name: Restore cache
-        uses: actions/cache@v3
-        with:
-          path: |
-            .next/cache
-          # Generate a new cache whenever packages or source files change.
-          key: ${{ runner.os }}-nextjs-${{ hashFiles('**/package-lock.json') }}-${{ hashFiles('**.[jt]s', '**.tsx') }}
-          # If source files changed but packages didn't, rebuild from a prior cache.
-          restore-keys: |
-            ${{ runner.os }}-nextjs-${{ hashFiles('**/package-lock.json') }}-
-      - name: Install dependencies
-        run: npm install
-      - name: Build with Next.js
-        run: npx --no-install next build
-      - name: Upload artifact
-        uses: actions/upload-artifact@v3
-        with:
-          path: 'out'
-          name: scra
+  #   steps:
+  #     - name: Checkout SCRA
+  #       uses: actions/checkout@v3
+  #       with:
+  #         repository: ASoftGlow/scra
+  #         token: ${{ secrets.GH_SCRA_PAT }}
+  #     - name: Setup Node
+  #       uses: actions/setup-node@v3
+  #       with:
+  #         node-version: "lts/*"
+  #         cache: npm
+  #     - name: Restore cache
+  #       uses: actions/cache@v3
+  #       with:
+  #         path: |
+  #           .next/cache
+  #         # Generate a new cache whenever packages or source files change.
+  #         key: ${{ runner.os }}-nextjs-${{ hashFiles('**/package-lock.json') }}-${{ hashFiles('**.[jt]s', '**.tsx') }}
+  #         # If source files changed but packages didn't, rebuild from a prior cache.
+  #         restore-keys: |
+  #           ${{ runner.os }}-nextjs-${{ hashFiles('**/package-lock.json') }}-
+  #     - name: Install dependencies
+  #       run: npm install
+  #     - name: Build with Next.js
+  #       run: npx --no-install next build
+  #     - name: Upload artifact
+  #       uses: actions/upload-artifact@v3
+  #       with:
+  #         path: 'out'
+  #         name: scra
           
   deploy:
     environment:
       name: github-pages
       url: ${{ steps.deployment.outputs.page_url }}
     runs-on: ubuntu-latest
-    needs: [build-dev, build-scra]
+    needs: [build-dev]
     steps:
       - name: Download dev artifact
         uses: actions/download-artifact@v3
         with:
           name: dev
-      - name: Download scra artifact
-        uses: actions/download-artifact@v3
-        with:
-          name: scra
-          path: scra
+      # - name: Download scra artifact
+      #   uses: actions/download-artifact@v3
+      #   with:
+      #     name: scra
+      #     path: scra
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v1
-        with:
-          # Upload entire repository
-          path: '.'
-      - name: Setup Pages
-        uses: actions/configure-pages@v3
+
       - name: Deploy to GitHub Pages
         id: deployment
         uses: actions/deploy-pages@v2
