+++ .gitignore
@@ -1,2 +1,2 @@
-.vscode
-css
+_site
+node_modules
