+++ _includes/footer.liquid
@@ -1,0 +1,4 @@
+<footer>
+    <span>© 2024 ASoftGlow</span>
+    <a href="mailto:me@asoftglow.dev">✉ Contact</a>
+</footer>
