+++ _includes/head.liquid
@@ -1,0 +1,15 @@
+<meta charset="utf-8">
+<meta name="robots" content="index, follow">
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<meta name="referrer" content="no-referrer">
+
+<title>ASoftGlow.dev{% if title %} | {{ title }}{% endif %}</title>
+<meta name="author" content="ASoftGlow">
+<meta name="creator" content="ASoftGlow">
+<meta name="description" content="{% if description %}{{ description }}{% else %}ASoftGlow's developer portfolio.{% endif %}">
+
+<meta name="theme-color" media="(prefers-color-scheme: light)" content="hsl(130, 33%, 66%)" />
+<meta name="theme-color" media="(prefers-color-scheme: dark)" content="hsl(130, 20%, 33%)" />
+
+<link rel="shortcut icon" href="/images/favicon.png">
+<link rel="stylesheet" href="/css/styles.css">
