+++ cat-generator.html
@@ -1,40 +1,24 @@
+---
+title: Cat generator
+description: Pixel art image generator
+---
 <!DOCTYPE html>
 <html lang="en-US">
 
 <head>
-    <meta charset="utf-8">
-    <meta name="robots" content="index, follow">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <meta name="referrer" content="no-referrer">
-
-    <title>ASoftGlow.dev</title>
-    <meta name="author" content="ASoftGlow">
-    <meta name="creator" content="ASoftGlow">
-    <meta name="description" content="ASoftGlow's developer portfolio.">
-
-    <meta name="theme-color" media="(prefers-color-scheme: light)" content="hsl(130, 33%, 66%)" />
-    <meta name="theme-color" media="(prefers-color-scheme: dark)" content="hsl(130, 20%, 33%)" />
-
-    <link rel="shortcut icon" href="images/favicon.png">
-    <link rel="stylesheet" href="css/styles.css">
-    <link rel="stylesheet" href="css/cat-generator.css">
+    {% include 'head' %}
+    <link rel="stylesheet" href="/css/cat-generator.css">
 </head>
 
 <body>
     <main>
-        <a id="home-link" href="/" title="Back">←</a>
-        <header>
-            <h2>Cat pixel art generator</h2>
-        </header>
+        {% include 'project-header' %}
         <div>
             <img width="50px" height="50px" id="cat-img" src="http://localhost:8080">
             <button id="new-cat-btn">New</button>
         </div>
     </main>
-    <footer>
-        <span>© 2024 ASoftGlow</span>
-        <a href="mailto:me@asoftglow.dev">✉ Contact</a>
-    </footer>
+    {% include 'footer' %}
 </body>
 
 </html>
