+++ css/index.scss
@@ -29,3 +29,7 @@
         margin-bottom: 30px;
     }
 }
+
+.disabled-a {
+    cursor: not-allowed;
+}
