+++ hockey.html
@@ -1,23 +1,13 @@
+---
+title: Hockey Line Picker
+description: Web app
+---
 <!DOCTYPE html>
 <html lang="en-US">
 
 <head>
-    <meta charset="utf-8">
-    <meta name="robots" content="index, follow">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <meta name="referrer" content="no-referrer">
-
-    <title>ASoftGlow.dev</title>
-    <meta name="author" content="ASoftGlow">
-    <meta name="creator" content="ASoftGlow">
-    <meta name="description" content="ASoftGlow's developer portfolio.">
-
-    <meta name="theme-color" media="(prefers-color-scheme: light)" content="hsl(130, 33%, 66%)" />
-    <meta name="theme-color" media="(prefers-color-scheme: dark)" content="hsl(130, 20%, 33%)" />
-
-    <link rel="shortcut icon" href="images/favicon.png">
-    <link rel="stylesheet" href="css/styles.css">
-    <link rel="stylesheet" href="css/hockey.css">
+    {% include 'head' %}
+    <link rel="stylesheet" href="/css/hockey.css">
 </head>
 
 <body>
@@ -26,20 +16,20 @@
             <a id="home-link" href="/" title="Back">←</a>
             <header>
                 <div>
-                    <h1>Hockey Line Picker</h1>
+                    <h1></h1>
                     <h3>App</h3>
-                    <a class="product-link" href="hlp">Link to product</a>
+                    <a class="product-link" href="/hlp">Link to product</a>
                 </div>
                 <div class="profile" id="icon" role="img"></div>
             </header>
 
             <p>A simple webapp for picking hockey lineups.</p>
             <figure>
-                <img src="images/hockey/menu.png" width="250px">
+                <img src="/images/hockey/menu.png" width="250px">
                 <figcaption>Select screen</figcaption>
             </figure>
             <figure>
-                <img src="images/hockey/display.png" width="250px">
+                <img src="/images/hockey/display.png" width="250px">
                 <figcaption>Display screen</figcaption>
             </figure>
             <section>
@@ -58,8 +48,8 @@
                 <h2>Original project</h2>
                 <figure>
                     <div class="gallery">
-                        <img src="images/hockey/old_menu.png">
-                        <img src="images/hockey/old_display.png">
+                        <img src="/images/hockey/old_menu.png">
+                        <img src="/images/hockey/old_display.png">
                     </div>
                     <figcaption>Select and display screens of Code.org project</figcaption>
                 </figure>
@@ -86,7 +76,7 @@
                 <p>
                     In the original project, players were hard coded. That's not ideal.
                     I added an add and remove player button to rectify that. Additionally, I gave the option to import players from a newline seperated text file, as to make it
-                    easier to add a bunch of players. <a download href="assets/players.txt">Download example file</a><br>
+                    easier to add a bunch of players. <a download href="/assets/players.txt">Download example file</a><br>
                     Unlike the original, the players and their selected positions are saved between refreshes using <a
                         href="https://developer.mozilla.org/en-US/docs/Web/API/Window/localStorage">localStorage</a>.
                 </p>
@@ -96,7 +86,7 @@
                     <a href="https://web.dev/learn/pwa/">Progressive Web App</a>.
                 </p>
                 <figure>
-                    <img src="images/hockey/display.png" width="250px">
+                    <img src="/images/hockey/display.png" width="250px">
                     <figcaption>New display screen</figcaption>
                 </figure>
                 <h3>Practice</h3>
@@ -106,17 +96,14 @@
                     Choose the players for each team and optionally a team color.
                 </p>
                 <figure>
-                    <img src="images/hockey/practice.png" width="250px">
+                    <img src="/images/hockey/practice.png" width="250px">
                     <figcaption>Practice page. It looks different than the other screenshots because FireFox was used.</figcaption>
                 </figure>
                 <ins>Note: the selections on this page are not saved and do not affect the main page.</ins>
             </section>
         </main>
     </div>
-    <footer>
-        <span>© 2024 ASoftGlow</span>
-        <a href="mailto:me@asoftglow.dev">✉ Contact</a>
-    </footer>
+    {% include 'footer' %}
 </body>
 
 </html>
