+++ menyow.html
@@ -1,35 +1,18 @@
+---
+title: Menyow
+description: Discord bot
+---
 <!DOCTYPE html>
 <html lang="en-US">
 
 <head>
-    <meta charset="utf-8">
-    <meta name="robots" content="index, follow">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <meta name="referrer" content="no-referrer">
-
-    <title>ASoftGlow.dev</title>
-    <meta name="author" content="ASoftGlow">
-    <meta name="creator" content="ASoftGlow">
-    <meta name="description" content="ASoftGlow's developer portfolio.">
-
-    <meta name="theme-color" media="(prefers-color-scheme: light)" content="hsl(130, 33%, 66%)" />
-    <meta name="theme-color" media="(prefers-color-scheme: dark)" content="hsl(130, 20%, 33%)" />
-
-    <link rel="shortcut icon" href="images/favicon.png">
-    <link rel="stylesheet" href="css/styles.css">
-    <link rel="stylesheet" href="css/menyow.css">
+    {% include 'head' %}
+    <link rel="stylesheet" href="/css/menyow.css">
 </head>
 
 <body>
     <main>
-        <a id="home-link" href="/" title="Back">←</a>
-        <header>
-            <div>
-                <h1>Menyow</h1>
-                <h3>Discord bot</h3>
-            </div>
-            <div class="profile" id="bot-icon" role="img"></div>
-        </header>
+        {% include 'project-header', image: 'https://cdn.discordapp.com/app-icons/809180078079606825/5047e7c1d83641f76392ade8112b8312.png?size=64' %}
         <p>
             At some point, my friend group moved from <a href="https://steampowered.com">Steam</a> to <a href="https://discord.com">Discord</a> for communication.
             Before then, I had never used Discord. One advantage it had over Steam was bots.
@@ -74,10 +57,7 @@
             </details>
         </div>
     </main>
-    <footer>
-        <span>© 2024 ASoftGlow</span>
-        <a href="mailto:me@asoftglow.dev">✉ Contact</a>
-    </footer>
+    {% include 'footer' %}
 </body>
 
 </html>
