+++ terrain.html
@@ -1,44 +1,28 @@
+---
+title: Terrain Generator
+---
 <!DOCTYPE html>
 <html lang="en-US">
 
 <head>
-    <meta charset="utf-8">
-    <meta name="robots" content="index, follow">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <meta name="referrer" content="no-referrer">
-
-    <title>ASoftGlow.dev</title>
-    <meta name="author" content="ASoftGlow">
-    <meta name="creator" content="ASoftGlow">
-    <meta name="description" content="ASoftGlow's developer portfolio.">
-
-    <meta name="theme-color" media="(prefers-color-scheme: light)" content="hsl(130, 33%, 66%)" />
-    <meta name="theme-color" media="(prefers-color-scheme: dark)" content="hsl(130, 20%, 33%)" />
-
-    <link rel="shortcut icon" href="images/favicon.png">
-    <link rel="stylesheet" href="css/styles.css">
-    <link rel="stylesheet" href="css/terrain.css">
+    {% include 'head' %}
+    <link rel="stylesheet" href="/css/terrain.css">
 </head>
 
 <body>
     <main>
-        <a id="home-link" href="/" title="Back">←</a>
-        <header>
-            <div>
-                <h1>Terrain Generator</h1>
-            </div>
-        </header>
+        {% include 'project-header' %}
         <article>
             <div class="elevation carousel">
                 <figure>
-                    <img src="images/terrain.png">
+                    <img src="/images/terrain.png">
                 </figure>
                 <figure>
-                    <img src="images/save_browser.png" alt="Save Browser">
+                    <img src="/images/save_browser.png" alt="Save Browser">
                     <figcaption>Save browser</figcaption>
                 </figure>
                 <figure>
-                    <img src="images/import_save.png" alt="Import save">
+                    <img src="/images/import_save.png" alt="Import save">
                     <figcaption>Import save</figcaption>
                 </figure>
             </div>
@@ -110,7 +94,7 @@
                 <h2>Save system</h2>
 
                 <figure>
-                    <img width="32px" src="images/glowsave.ico" alt="glowsave icon" style="image-rendering: pixelated;">
+                    <img width="32px" src="/images/glowsave.ico" alt="glowsave icon" style="image-rendering: pixelated;">
                     <figcaption style="display: inline-block;"><code>.glowsave</code> file type icon</figcaption>
                 </figure>
                 <details open style="position: relative;">
@@ -170,7 +154,7 @@
             <section>
                 <h2>Generation</h2>
                 <figure>
-                    <img src="images/mesh.png" alt="Triangles mesh view">
+                    <img src="/images/mesh.png" alt="Triangles mesh view">
                     <figcaption>Debug line view, showing triangles</figcaption>
                 </figure>
                 <h3>Issues</h3>
@@ -181,10 +165,7 @@
             </section>
         </article>
     </main>
-    <footer>
-        <span>© 2024 ASoftGlow</span>
-        <a href="mailto:me@asoftglow.dev">✉ Contact</a>
-    </footer>
+    {% include 'footer' %}
 </body>
 
 </html>
